// src/lib/generated/kaizenAdmin/models/approvalLevelApprovalType.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * KaizenAdmin Approval System
 * KaizenAdmin and Approval Management System using LangGraph
 * OpenAPI spec version: 1.0.0
 */

/**
 * Type of approval - by role or specific individual
 */
export type ApprovalLevelApprovalType = typeof ApprovalLevelApprovalType[keyof typeof ApprovalLevelApprovalType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ApprovalLevelApprovalType = {
  role: 'role',
  individual: 'individual',
} as const;